fixing secret data reactor
