+++ action/report_test.go
@@ -45,9 +45,7 @@ func TestKubeReportText(t *testing.T) {
 			Name:      "default-1-spec",
 			Namespace: "default",
 		},
-		Data: map[string][]byte{
-			"experiment.yaml": byteArray,
-		},
+		StringData: map[string]string{"experiment.yaml": string(byteArray)},
 	}, metav1.CreateOptions{})
 
 	byteArray, _ = ioutil.ReadFile(base.CompletePath("../testdata/assertinputs", "result.yaml"))
@@ -56,9 +54,7 @@ func TestKubeReportText(t *testing.T) {
 			Name:      "default-1-result",
 			Namespace: "default",
 		},
-		Data: map[string][]byte{
-			"result.yaml": byteArray,
-		},
+		StringData: map[string]string{"result.yaml": string(byteArray)},
 	}, metav1.CreateOptions{})
 
 	err := rOpts.KubeRun(os.Stdout)
